% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_control.R
\name{cipres_delete_job}
\alias{cipres_delete_job}
\title{Delete or cancel jobs on CIPRES server}
\usage{
cipres_delete_job(handle, verbose = TRUE, ...)
}
\arguments{
\item{handle}{The job identifier}

\item{verbose}{Should a message be printed if the
deletion/cancellation is successful?}

\item{...}{arguments to customize the API call. These arguments are passed to the appropriate \code{httr} function (\code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{DELETE}})}
}
\value{
\code{TRUE} if the deletion/cancellation was succesfull,
    \code{FALSE} otherise.
}
\description{
Deletes data and output associated with a job on CIPRES's server
(or cancel a job if it's still running.)
}
\details{
Given a job handle, deletes data and output associated with the
job. If the job hasn't completed, it will abort it and deletes the
data and output.
}
\examples{
\dontrun{
### Deletes the 5 oldest jobs. Be careful, there is no undo!
### sapply(cipres_list_jobs()$title[1:5], cipres_delete_jobs)
}
}
\seealso{
Other job.controls: \code{\link{cipres_job_status}},
  \code{\link{cipres_list_jobs}}
}

