% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_control.R
\name{cipres_job_status}
\alias{cipres_job_status}
\title{Job status}
\usage{
cipres_job_status(handle, ...)
}
\arguments{
\item{handle}{The job identifier}

\item{...}{arguments to customize the API call. These arguments are passed to the appropriate \code{httr} function (\code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{DELETE}})}
}
\value{
A list containing the job handle, the job status
    (\code{stage}), whether the job failed, and the job data
    submission.
}
\description{
Check the status of a job.
}
\examples{
\dontrun{
  ## Status of the latest job submitted
  lst_jobs <- cipres_list_jobs()
  cipres_job_status(lst_jobs$handle[nrow(lst_jobs)])
}
}
\author{
Francois Michonneau
}
\seealso{
Other job.controls: \code{\link{cipres_delete_job}},
  \code{\link{cipres_list_jobs}}
}

