% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{cipres_list_files}
\alias{cipres_download}
\alias{cipres_list_files}
\title{List and download files associated with a job}
\usage{
cipres_list_files(job_handle, ...)

cipres_download(job_handle, outdir, pattern = NULL, ignore.case = FALSE,
  show_progress = TRUE, ...)
}
\arguments{
\item{job_handle}{the job identified}

\item{...}{arguments to customize the API call. These arguments are passed to the appropriate \code{httr} function (\code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{DELETE}})}

\item{outdir}{the path where the files should be downloaded}

\item{pattern}{an optional regular expression to download only the
files that match the pattern}

\item{ignore.case}{should pattern matching be case sensitive?}

\item{show_progress}{should the progress of the files being
downloaded be displayed?}
}
\value{
\code{cipres_list_files} returns a data frame listing all
    the files associated with a job. \code{cipres_download_file}
    and \code{cipres_download_all} returns whether the files were
    successfully downloaded.
}
\description{
Find and download the files associated with a job.
}
\details{
Be careful, files with identical names will be silently overwritten.

By only providing a job handle to \code{cipres_download}, all the
files generated by the job will be downloaded. To download only
one or a subset of these files, you can use regular expression
pattern matching.
}
\examples{
\dontrun{
### List the files associated with the oldest job
### cipres_list_files(cipres_list_jobs$handle[1])

### Download all the files associated with the oldest job
### cipres_download(cipres_list_jobs$handle[1], tempdir())
}
}

