% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{cipres_list_jobs}
\alias{cipres_list_jobs}
\title{List jobs currently running on CIPRES}
\usage{
cipres_list_jobs()
}
\value{
XML
}
\description{
The list of jobs currently running on CIPRES
}
\author{
Francois Michonneau
}

