% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cipres_login.R
\name{cipres_login}
\alias{cipres_login}
\title{Access login information for the CIPRES API}
\usage{
cipres_login()
}
\value{
A list with the login information (user, password, and
    application ID).
}
\description{
List with the login information
}
\details{
This function is mostly intended to be called internally but could
be useful to verify the login information.
}
\author{
Francois Michonneau
}

