% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cipres_login.R
\name{cipres_setup}
\alias{cipres_setup}
\title{Set up CIPRES login information}
\usage{
cipres_setup()
}
\description{
Interactive function to setup the credential information
}
\details{
This function guides you through the steps needed to store your
login information for using the CIPRES API. You should use it the
first time you use this package.
}
\author{
Francois Michonneau
}

