% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{cipres_submit_job}
\alias{cipres_submit_job}
\title{Submit a job to CIPRES}
\usage{
cipres_submit_job(tool = "RAXMLHPC8_REST_XSEDE",
  input_file = "./inst/raxml_inputphy.txt")
}
\arguments{
\item{tool}{The tool to use to analyze the data}

\item{input_file}{The input file to use with the selected tool}
}
\value{
XML
}
\description{
Submit a job to CIPRES using the API
}
\details{
Possible tools are ...
}
\author{
Francois Michonneau
}

