% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raxml.R
\name{cipres_submit_raxml}
\alias{cipres_submit_raxml}
\title{Submit RAxML analysis}
\usage{
cipres_submit_raxml(input_file, select_analysis = c("fa", "fd", "fo", "fA",
  "J", "y", "I"), starting_tree = NULL, constraint = NULL,
  binary_backbone = NULL, partition = NULL, exclude = NULL,
  weights = NULL, user_prot_matrix = NULL, secondary_structure = NULL,
  bootstrap_topologies = NULL, max_runtime = 10, CAT_categories = NULL,
  parsimony_seed = NULL, datatype = c("dna", "protein", "rna", "binary",
  "mutli"), outgroups = NULL, p_inv = FALSE, empirical_freq = TRUE,
  print_brlen = FALSE, more_memory = FALSE, n_char = NULL, n_tax = NULL,
  disable_check_sequence = FALSE, mesquite_output = FALSE,
  model = "GTRGAMMA", n_runs = 1, no_bfgs = FALSE,
  intermediate_tree_files = FALSE, convergence_criterion = FALSE,
  majority_rule = c("MRE", "STRICT", "MR_DROP", "STRICT_DROP"),
  protein_matrix = c("DAYHOFF", "DCMUT", "JTT", "MTREV", "WAG", "RTREV",
  "CPREV", "VT", "BLOSUM62", "MTMAM", "LG", "MTART", "MTZOA", "PMB", "HIVB",
  "HIVW", "JTTDCMUT", "FLU", "DUMMY", "DUMMY2", "AUTO", "LG4M", "LG4X",
  "PROT_FILE", "GTR_UNLINKED", "GTR"), bootstrap_type = c("b", "x"),
  bs_seed_value = 12345, n_bootstrap_rep = NULL,
  bootstop_type = c("autoMRE", "autoFC", "autoMR", "autoMRE_IGN"),
  job_name = NULL, get_email = TRUE, ...)
}
\arguments{
\item{input_file}{An alignment file in relaxed interleaved for
sequential PHYLIP (or FASTA) format (-s)}

\item{select_analysis}{The type of analysis to be performed.}

\item{starting_tree}{A NEWICK tree file specifying the starting
tree. If \code{NULL} (default), RAxML uses a randomized
stepwise addition Parsimony starting tree.}

\item{constraint}{A NEWICK tree file specifying topology
constraining the tree search. If \code{NULL} (default), no
constraint is used.}

\item{binary_backbone}{The file name of a binary constraint
tree. If \code{NULL} (default), no constraint is used.}

\item{partition}{A file specifying the partitions in the alignment
file using the RAxML format (refer to the RAxML manual for
details). If \code{NULL} (default), the full alignment is
considered as a single partition.}

\item{exclude}{A file specifying positions in the alignment to
exclude. If \code{NULL} (default), no position is excluded.}

\item{weights}{A file indicating the weights for each column
position in the alignment. If \code{NULL} (default), all
alignments positions have identical weights.}

\item{user_prot_matrix}{A file containing an amino-acid
substitution model.}

\item{secondary_structure}{A file specifying a secondary structure
model for the alignment.}

\item{bootstrap_topologies}{A file containing topologies for a
posteriori bootstrapping.}

\item{max_runtime}{How long should the job run for? (in hours)}

\item{CAT_categories}{When using a CAT model, how many rate
categories (-c)?}

\item{parsimony_seed}{For all options that use randomization, this
seed must be specified (-p).}

\item{datatype}{The type of data used in the alignment.}

\item{outgroups}{Specify in name of a single outgroup
(\dQuote{Cat}) or a comma-separated list of outgroups (no
spaces: \dQuote{Cat,Dog,Bird}). If several species are listed
as outgroup, only the first one in the list will be used a
outgroup.}

\item{p_inv}{Should a proportion of invariant site be estimated?
(default \code{FALSE}).}

\item{empirical_freq}{Should empirical frequencies be used with
amino-acid alignments? (default \code{FALSE}).}

\item{print_brlen}{Should branch lengths be printed on the screen?
Default \code{FALSE}.}

\item{more_memory}{Set to \code{TRUE} if your analysis requires
more than 125GB of memory.}

\item{n_char}{If using \code{more_memory}, set to the number of
characters in your alignment to estimate the amount of memory
needed.}

\item{n_tax}{If using \code{more_memory}, set to the number of
taxa in your alignment to estimate the amount of memory
needed.}

\item{disable_check_sequence}{Disable checking for completely
undetermined sequences (it normally doesn't make sense to
incude them in the analysis). Default
\code{FALSE}. (\code{-O}).}

\item{mesquite_output}{Generate output files that can be parsed by
MESQUITE. Default \code{FALSE}. (Not allowed by CIPRES in runs
with bootstrapping or alternative runs are used)
(\code{-mesquite}).}

\item{model}{The substitution model (\code{-m}).}

\item{n_runs}{The number of alternative runs on distinct starting
trees (\code{-#} or \code{-N}).}

\item{no_bfgs}{If \code{TRUE}, disable the BFGS method when
optimizing the GTR rates (default \code{FALSE}).}

\item{intermediate_tree_files}{If \code{TRUE}, the intermediate
tree will be written to files (\code{-j}).}

\item{convergence_criterion}{ML search convergence criterion uses
Robinson-Foulds distance. Default \code{FALSE}. Use on trees
with large number of taxa (\code{-D}).}

\item{majority_rule}{When computing the majority rule consensus
tree (\code{select_analysis="J"}), which method to use?}

\item{protein_matrix}{Which amino-acid substitution matrix to use
when working with protein sequence alignment?}

\item{bootstrap_type}{Which type of bootstrapping algorithm to use
(with \code{select_analysis} is \code{fa}, \code{fd}, or
\code{fo})? \code{x} for rapid bootstrapping and \code{b} for
normal bootstrapping.}

\item{bs_seed_value}{The seed to use for the bootstrapping.}

\item{n_bootstrap_rep}{If \code{NULL}, the number of bootstrap
replicate will be selected by RAxML according to the algorithm
specified by the option \code{bootstop_type}. Other an integer
specifiying the number of replicates.}

\item{bootstop_type}{The type of algorithm used to estimate the
number of replicates.}

\item{job_name}{An optional name to identify the job.}

\item{get_email}{Should an email be sent when the job completed?}

\item{...}{arguments to customize the API call. These arguments are passed to the appropriate \code{httr} function (\code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{DELETE}})}
}
\description{
Submit to CIPRES an RAxML job. For more information, please
consult the RAxML manual available on the RAxML website
(\url{http://sco.h-its.org/exelixis/web/software/raxml}).
}
\details{
CIPRES offers several flavors of RAxML but \code{cipresr} only
provides an interface with the full RAxML version. Setting the
correct options for a given run can be challenging as RAxML is
complex and feature-rich software. Read the RAxML manual
carefully, double check for potential warnings and error messages
in the output files, and your results.
}
\author{
Francois Michonneau
}

