% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{cipres_list_files}
\alias{cipres_download_all}
\alias{cipres_download_file}
\alias{cipres_list_files}
\title{List and download files associated with a job}
\usage{
cipres_list_files(job_handle, ...)

cipres_download_file(job_handle, file_id, outfile, ...)

cipres_download_all(job_handle, outdir, pattern = NULL, ignore.case = FALSE,
  ...)
}
\arguments{
\item{job_handle}{the job identified}

\item{...}{arguments to customize the API call. These arguments are passed to the appropriate \code{httr} function (\code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{DELETE}})}

\item{file_id}{the file identifier (outputDocumentId, not the file name)}

\item{outfile}{the path and file name where the output should be written}

\item{outdir}{the path where the files should be downloaded}

\item{pattern}{an optional regular expression to download only the
files that match the pattern}

\item{ignore.case}{should pattern matching be case sensitive?}
}
\value{
\code{cipres_list_files} returns a data frame listing all
    the files associated with a job. \code{cipres_download_file}
    and \code{cipres_download_all} returns whether the files were
    successfully downloaded.
}
\description{
Be careful, files with identical names will be silently overwritten.
}
\examples{
\dontrun{
### List the files associated with the oldest job
### cipres_list_files(cipres_list_jobs$handle[1])

### Download all the files associated with the oldest job
### cipres_download_all(cipres_list_jobs$handle[1], tempdir())
}
}

