% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_control.R
\name{cipres_list_jobs}
\alias{cipres_list_jobs}
\title{List past and current jobs on the CIPRES server}
\usage{
cipres_list_jobs(...)
}
\arguments{
\item{...}{arguments to customize the API call. These arguments are passed to the appropriate \code{httr} function (\code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{DELETE}})}
}
\value{
A data frame with the job handle and its URL
}
\description{
List past and running jobs on the CIPRES server.
}
\seealso{
Other job.controls: \code{\link{cipres_delete_job}},
  \code{\link{cipres_job_status}}
}

