% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{cipres_login}
\alias{cipres_login}
\title{Access login information for the CIPRES API}
\usage{
cipres_login(user = Sys.getenv("CIPRES_USER"),
  password = Sys.getenv("CIPRES_PWD"), app_id = Sys.getenv("CIPRES_APP_ID"))
}
\arguments{
\item{user}{Your CIPRES API user name}

\item{password}{Your CIPRES API password}

\item{app_id}{Your CIPRES API application token}
}
\value{
A list
}
\description{
Easy programmatic access to the login information. This function
shouldn't be used by normal users and is only intended to be
called by other functions.
}
\details{
You need to create the environment variables \code{CIPRES_USER},
\code{CIPRES_PWD}, \code{CIPRES_APP_ID} in your .Renviron file
after registering for an account at
\url{https://www.phylo.org/restusers/login.action}. Therefore, you
will need to enter your login information only once.
}
\author{
Francois Michonneau
}

